cleanup and code deduplication
* ClientPattern.cc: make compiler happy (complaints about XPROP not handled in switch)
* FbTk/FbPixmap.cc FbTk/StringUtil.cc: signed vs unsigned
* FbTk/FbString.cc: missing return
* WinClient.cc: create atoms only once; use helper function
* Window.cc: use a helper function

+++ src/FbTk/FbPixmap.cc
@@ -369,7 +369,7 @@ bool FbPixmap::rootwinPropertyNotify(int screen_num, Atom atom) {
         return false;
 
     checkAtoms();
-    for (int i=0; i < sizeof(root_props)/sizeof(RootProps); ++i) {
+    for (size_t i = 0; i < sizeof(root_props)/sizeof(RootProps); ++i) {
         if (root_props[i].atom == atom) {
             Pixmap root_pm = None;
             Atom real_type;
